Expanded Bag

+++ include/item.h
@@ -24,7 +24,7 @@ struct Item
 struct BagPocket
 {
     struct ItemSlot *itemSlots;
-    u8 capacity;
+    u16 capacity;
 };
 
 extern struct BagPocket gBagPockets[];
